requirements.txt
numpy==1.23.4
pandas==1.5.2
Pillow==9.3.0
progress==1.6
scikit_learn==1.1.3
semilearn==0.2.4
scikit-image
tensorboard
torch
torchvision
